DEV:  integrate root_to_dev@1766
[git-p4: depot-paths = "//open/dev/": change = 1767]

+++ farrago/initBuild.sh
@@ -3,18 +3,32 @@
 
 # Script to set up a new Farrago build environment
 
+usage() {
+    echo "Usage:  initBuild.sh --with[out]-fennel [--append-init-properties]"
+}
+
 if [ "$1" == "--with-fennel" ] ; then
     fennel_disabled=false
 elif [ "$1" == "--without-fennel" ] ; then
     fennel_disabled=true
 else
-    echo "Usage:  initBuild.sh [--with-fennel|--without-fennel]"
+    usage
     exit -1
 fi
-    
+
+
+if [ "$2" == "--append-init-properties" ] ; then
+    touch initBuild.properties
+elif [ -z "$2" ] ; then
+    # default is remove this file
+    rm -f initBuild.properties
+else
+    usage
+    exit -1
+fi
+
 # Set up Farrago custom build properties file
-rm -f initBuild.properties
-cat > initBuild.properties <<EOF
+cat >> initBuild.properties <<EOF
 # initBuild.properties should only be used to store the fennel.disabled
 # property: initBuild.sh will destroy other information stored here.  Create
 # customBuild.properties to override other build parameters if necessary.
